Rename synapse.async for python 3.7 compat
